skims\test(back): #4331.21 add functional test
- add test to validate the reference between inst instances
- some classes cannot be found


Former-commit-id: 905131031a4a29328fb2b94968dedaafd47247cd

+++ skims/test/data/benchmark/instance_references/src/App.java
@@ -29,11 +29,11 @@ public class App {
     }
 
     public void test_02(){
-        int rand = new java.util.Random().nextFloat();
+        int rand = new java.util.Random().nextInt();
 
         HttpServletRequest request = new HttpServletRequest();
 
-        User currentUser = new User("Jane", "Doe", Float.toString(rand));
+        User currentUser = new User("Jane", "Doe", Integer.toString(rand));
 
         request.getSession().setAttribute("testInstanceReference", currentUser.lastName);
     }
